#49 Merge

+++ Vagrantfile
@@ -38,7 +38,7 @@ SCRIPT
 
 $solr = <<SCRIPT
 cd /home/vagrant
-mkdir "downloads"
+mkdir -p "downloads"
 cd downloads
 SOLR_TAR="solr-7.7.2.tgz"
 if test ! -f "$SOLR_TAR"; then
@@ -49,18 +49,23 @@ cd /home/vagrant/solr-7.7.2
 mkdir -p server/solr/opus4/conf
 echo name=opus4 > server/solr/opus4/core.properties
 cd server/solr/opus4/conf/
-ln -s /vagrant/conf/schema.xml schema.xml
-ln -s /vagrant/conf/solrconfig.xml solrconfig.xml
+if test ! -f "schema.xml"; then
+  ln -s /vagrant/conf/schema.xml schema.xml
+fi
+if test ! -f "solrconfig.xml"; then
+  ln -s /vagrant/conf/solrconfig.xml solrconfig.xml
+fi
 SCRIPT
 
 $database = <<SCRIPT
-export MYSQL_PWD=root && mysql --default-character-set=utf8 -h 'localhost' -P '3306' -u 'root' -v -e "CREATE DATABASE IF NOT EXISTS opusdb DEFAULT CHARACTER SET = UTF8 DEFAULT COLLATE = UTF8_GENERAL_CI; CREATE USER IF NOT EXISTS 'opus4admin'@'localhost' IDENTIFIED WITH mysql_native_password BY 'root'; GRANT ALL PRIVILEGES ON opusdb.* TO 'opus4admin'@'localhost'; CREATE USER IF NOT EXISTS 'opus4'@'localhost' IDENTIFIED WITH mysql_native_password BY 'root'; GRANT SELECT,INSERT,UPDATE,DELETE ON opusdb.* TO 'opus4'@'localhost'; FLUSH PRIVILEGES;"
+/vagrant/vendor/opus4-repo/framework/bin/prepare-database.sh --admin_pwd root --user_pwd root
 SCRIPT
 
 $opus = <<SCRIPT
 cd /vagrant
 ant prepare-workspace prepare-config -DdbUserPassword=root -DdbAdminPassword=root
-php test/TestAsset/createdb.php
+export APPLICATION_PATH=/vagrant
+php vendor/opus4-repo/framework/db/createdb.php
 SCRIPT
 
 $environment = <<SCRIPT
